LayoutTests/accessibility/mac/tab-focus-post-notification-expected.txt

This tests that tabbing focus onto elements will send out notifications

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS count == 0 is true
PASS successfullyParsed is true

TEST COMPLETE
